+++ assets/js/blocks/reviews-by-product/style.scss
@@ -1,4 +1,38 @@
 .wc-block-reviews-by-product {
+	.is-loading {
+		.wc-block-reviews-by-product__text {
+			@include placeholder();
+			display: block;
+			width: 60%;
+		}
+
+		.wc-block-reviews-by-product__info {
+			.wc-block-reviews-by-product__avatar {
+				@include placeholder();
+			}
+
+			.wc-block-reviews-by-product__meta {
+				.wc-block-reviews-by-product__author {
+					@include placeholder();
+					font-size: 1em;
+					width: 80px;
+				}
+
+				.wc-block-reviews-by-product__rating {
+					.wc-block-reviews-by-product__rating__stars > span {
+						display: none;
+					}
+				}
+			}
+
+			.wc-block-reviews-by-product__published-date {
+				@include placeholder();
+				height: 1em;
+				width: 120px;
+			}
+		}
+	}
+
 	.wc-block-reviews-by-product__list {
 		list-style: none;
 		margin: 0;
@@ -11,20 +45,21 @@
 	.wc-block-reviews-by-product__info {
 		display: grid;
 		grid-template-columns: 1fr;
+		grid-template-rows: 1fr 0.65fr;
 		margin-top: $gap-small;
 
-		> .avatar {
+		.wc-block-reviews-by-product__avatar {
+			height: 48px;
 			grid-column: 1;
 			grid-row: 1 / 3;
+			width: 48px;
 		}
 
 		.wc-block-reviews-by-product__meta {
+			align-items: center;
+			display: flex;
 			grid-column: 1;
 			grid-row: 1;
-
-			.wc-block-reviews-by-product__author {
-				font-size: 1.4em;
-			}
 
 			.wc-block-reviews-by-product__author + .wc-block-reviews-by-product__rating {
 				margin-left: $gap-smaller;
@@ -32,10 +67,11 @@
 
 			.wc-block-reviews-by-product__rating {
 				display: inline;
+				height: 1em;
 
-				> .star-rating {
+				> .wc-block-reviews-by-product__rating__stars {
 					display: inline-block;
-					top: 5px;
+					top: -5px;
 					overflow: hidden;
 					position: relative;
 					height: 1.618em;
@@ -46,7 +82,7 @@
 					font-weight: 400;
 				}
 
-				> .star-rating::before {
+				> .wc-block-reviews-by-product__rating__stars::before {
 					content: '\53\53\53\53\53';
 					opacity: .25;
 					float: left;
@@ -55,7 +91,7 @@
 					position: absolute;
 				}
 
-				> .star-rating span {
+				> .wc-block-reviews-by-product__rating__stars span {
 					overflow: hidden;
 					float: left;
 					top: 0;
@@ -64,7 +100,7 @@
 					padding-top: 1.5em;
 				}
 
-				> .star-rating span::before {
+				> .wc-block-reviews-by-product__rating__stars span::before {
 					content: '\53\53\53\53\53';
 					top: 0;
 					position: absolute;
@@ -75,28 +111,29 @@
 		}
 
 		.wc-block-reviews-by-product__published-date {
+			align-self: center;
 			color: #808080;
-			font-size: 0.9em;
+			font-size: 0.875em;
 			grid-column: 1;
 			grid-row: 1;
 		}
 	}
 
-  &.has-picture {
-    .wc-block-reviews-by-product__info {
-  		grid-template-columns: #{60px + $gap-small} 1fr;
-    }
+	&.has-picture {
+		.wc-block-reviews-by-product__info {
+			grid-template-columns: #{48px + $gap-small} 1fr;
+		}
 
-    .wc-block-reviews-by-product__meta,
-    .wc-block-reviews-by-product__published-date {
-      grid-column: 2;
-    }
-  }
+		.wc-block-reviews-by-product__meta,
+		.wc-block-reviews-by-product__published-date {
+			grid-column: 2;
+		}
+	}
 
-  &.has-name,
-  &.has-rating {
-    .wc-block-reviews-by-product__published-date {
-      grid-row: 2;
-    }
-  }
+	&.has-name,
+	&.has-rating {
+		.wc-block-reviews-by-product__published-date {
+			grid-row: 2;
+		}
+	}
 }
